refactor: place issue/redeem/replace/refund request in primitives
